requirement.txt
absl-py==0.14.0
albumentations==1.0.3
appier==1.25.2
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
boto3==1.18.47
botocore==1.21.47
bravado==11.0.3
bravado-core==5.17.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cloudpickle==2.0.0
ctk-cli==1.5
cycler==0.10.0
Cython==0.29.24
dask==2021.12.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dgl-cu111==0.7.2
diskcache==5.4.0
distributed==2021.12.0
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
fonttools==4.33.2
fsspec==2021.11.1
future==0.18.2
gast==0.4.0
GDAL==3.5.0
gdown==3.13.1
girder-client==3.1.7
girder-slicer-cli-web==1.2.3
gitdb==4.0.7
GitPython==3.1.24
Glymur==0.9.7
google-auth==1.35.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.40.0
HeapDict==1.0.1
histomicstk==1.1.0
idna==3.2
imagecodecs==2021.11.20
imageio==2.9.0
imgaug==0.4.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
large-image==1.9.0
large-image-source-bioformats==1.9.0
large-image-source-deepzoom==1.9.0
large-image-source-dummy==1.9.0
large-image-source-gdal==1.9.0
large-image-source-mapnik==1.9.0
large-image-source-nd2==1.9.0
large-image-source-ometiff==1.9.0
large-image-source-openjpeg==1.9.0
large-image-source-openslide==1.9.0
large-image-source-pil==1.9.0
large-image-source-test==1.9.0
large-image-source-tiff==1.9.0
libtiff==0.5.0
llvmlite==0.37.0
locket==0.2.1
lxml==4.7.1
mapnik==4.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
msgpack==1.0.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nd2reader==3.3.0
neptune-client==0.16.15
nest-asyncio==1.5.1
networkx==2.6.3
nimfa==1.4.0
notebook==6.4.4
numba==0.54.0
numpy==1.19.5
oauthlib==3.1.1
onnx==1.10.2
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
openslide-python==1.1.2
packaging==21.0
palettable==3.3.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
Pillow==8.3.2
PIMS==0.5
pkgconfig==1.5.5
plotly==5.7.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.1
pycparser==2.20
pyflann==1.6.14
pyflann-py3==0.1.0
Pygments==2.10.0
PyJWT==2.1.0
pykeops==1.5
pynndescent==0.5.4
pynvml==11.0.0
pyparsing==2.4.7
pyproj==3.3.0.1
pyroomacoustics==0.5.0
pyrsistent==0.18.0
PySocks==1.7.1
python-bioformats==4.0.5
python-dateutil==2.8.2
python-javabridge==4.0.3
pytz==2021.1
pyvips==2.1.16
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
rdflib==6.1.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3987==1.3.8
rsa==4.7.2
s3transfer==0.5.0
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
Shapely==1.8.0
simplejson==3.17.5
six==1.16.0
sklearn==0.0
slack-api==0.1.4
slack-sdk==3.13.0
slicerator==1.0.0
smmap==4.0.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.29
strict-rfc3339==0.7
submitit==1.4.0
swagger-spec-validator==2.7.4
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thop @ git+https://github.com/Lyken17/pytorch-OpCounter.git@2f9bb08b1728666f8cbfb1c099f3c6b0d8c15889
threadpoolctl==2.2.0
tifffile==2021.8.30
tifftools==1.2.9
timm==0.4.12
toolz==0.11.2
torch==1.9.1+cu111
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.13
torch-spline-conv==1.2.1
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
umap==0.1.1
umap-learn==0.5.1
urllib3==1.26.7
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xmltodict==0.12.0
zict==2.0.0
